Make it run locally

+++ im_exec.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import tkinter
+from pathlib import Path
 
 # os.environ["MAGICK_HOME "] = os.getcwd() + R"\imageMagick"
 
@@ -21,8 +22,11 @@ def compress_dir_jpg(
         source_dir = source_dir.rstrip("\\")
         output_dir = output_dir.rstrip("\\")
         total_count = recr_struct_count_im(source_dir, output_dir, extension)
-        label_var.set("ROZPOCZĘTO")
-        label_current_no.set("0/" + str(total_count))
+        
+        if label_var is not None:
+            label_var.set("ROZPOCZĘTO")
+        if label_current_no is not None:
+            label_current_no.set("0/" + str(total_count))
         image_counter = 0
         return_code = 0
         if isinstance(extension, str):
@@ -51,15 +55,21 @@ def compress_dir_jpg(
     except Exception as e:
         print("Error: ", e)
         # if e is ChildProcessError:
-        error_var.set(e)
+        if error_var is not None:
+            error_var.set(e)
         # else:
         #     error_var.set("Coś poszło nie tak")
 
-    stop_btn_handle.configure(command=None)
-    stop_btn_handle.grid_forget()
-    label_var.set("ZAKOŃCZONO")
-    label_current_no.set("")
-    compress_btn_handle.configure(state=tkinter.NORMAL)
+    if stop_btn_handle is not None:
+        stop_btn_handle.configure(command=None)
+        stop_btn_handle.grid_forget()
+        
+    if label_var is not None:
+        label_var.set("ZAKOŃCZONO")
+    if label_current_no is not None:
+        label_current_no.set("")
+    if compress_btn_handle is not None:
+        compress_btn_handle.configure(state=tkinter.NORMAL)
 
 
 def compress_dir_proc(
@@ -75,27 +85,31 @@ def compress_dir_proc(
     error_var: tkinter.StringVar,
     extension: str = ".jpg",
 ):
-    cmd = (
-        os.getcwd()
-        + "\\imageMagick\\magick mogrify -path "
-        + '"'
-        + output_dir
-        + '"'
-        + " -resize 3840x2160^ -filter Triangle -define filter:support=2 -unsharp 0.25x0.08+8.3+0.045 -dither None "
-        + "-quality "
-        + str(quality)
-        + " -define jpeg:fancy-upsampling=off -define png:compression-filter=5 "
-        + "-define png:compression-level=9 -define png:compression-strategy=1 -interlace none -colorspace sRGB -clamp -verbose "
-        + '"'
-        + source_dir
-        + "\\*"
-        + extension
-        + '"'
-    )
+    magick_path = Path(os.getcwd()) / "imageMagick" / "magick.exe"
+    cmd = [
+        str(magick_path), "mogrify",
+        "-path", str(output_dir),
+        "-resize", "3840x2160^",
+        "-filter", "Triangle",
+        "-define", "filter:support=2",
+        "-unsharp", "0.25x0.08+8.3+0.045",
+        "-dither", "None",
+        "-quality", str(quality),
+        "-define", "jpeg:fancy-upsampling=off",
+        "-define", "png:compression-filter=5",
+        "-define", "png:compression-level=9",
+        "-define", "png:compression-strategy=1",
+        "-interlace", "none",
+        "-colorspace", "sRGB",
+        "-clamp",
+        "-verbose",
+        str(Path(source_dir) / f"*{extension}")
+    ]
     print(cmd)
     prev_filename = ""
     proc_im = execute(cmd)
-    stop_btn_handle.configure(command=lambda: proc_im.terminate())
+    if stop_btn_handle is not None:
+        stop_btn_handle.configure(command=lambda: proc_im.terminate())
     return_code = 0
 
     for line in read_process_output(proc_im, cmd):
@@ -105,14 +119,17 @@ def compress_dir_proc(
             break
         filename = extract_filename(line, extension)
         if filename != extension:
-            label_var.set(filename + extension)
+            if label_var is not None:
+                label_var.set(filename + extension)
             if filename != prev_filename:
                 prev_filename = filename
                 image_counter += 1
-                label_current_no.set(
+                if label_current_no is not None:
+                    label_current_no.set(
                     str(image_counter) + "/" + str(total_count)
                 )
-                progress_bar_fn(float(image_counter / total_count))
+                if progress_bar_fn is not None:
+                    progress_bar_fn(float(image_counter / total_count))
 
     return image_counter, return_code
 
@@ -124,7 +141,7 @@ def recr_struct_count_im(source_dir, output_dir, extension):
         if not os.path.isdir(structure):
             os.mkdir(structure)
         else:
-            print("Folder does already exits!")
+            print("Folder does already exits! " + structure)
         if isinstance(extension, str):
             extension = [extension]
 
@@ -186,14 +203,18 @@ if __name__ == "__main__":
     #     R"D:\OneDrive - Politechnika Wroclawska\projectsPython\imageCompressor\images_in\P5130202.JPG JPEG 4608x3456 4608x3456+0+0 8-bit sRGB 6.83107MiB 0.172u 0:00.170>3840x2880 3840x2880+0+0 8-bit sRGB 860718B 1.313u 0:01.319",
     #     ".jpg",
     # )
-    source_dir = R"C:\Users\micha\Pictures\compress"
-    output_dir = R"C:\Users\micha\Pictures\compressed"
-    print(recr_struct_count_im(source_dir, output_dir, ".jpg"))
+    source_dir = R"C:\Users\lapci\Pictures\magick_tests\test"
+    output_dir = R"C:\Users\lapci\Pictures\magick_tests\test_comp"
+    # print(recr_struct_count_im(source_dir, output_dir, ".jpg"))
     # print(calculate_images(source_dir, ".jpg"))
-    # compress_dir_jpg(
-    #     source_dir=source_dir,
-    #     output_dir=output_dir,
-    #     progress_bar_fn=None,
-    #     compress_btn_handle=None,
-    #     label_var=None,
-    # )
+    compress_dir_jpg(
+        source_dir=source_dir,
+        output_dir=output_dir,
+        progress_bar_fn=None,
+        compress_btn_handle=None,
+        label_var=None,
+        label_current_no=None,
+        stop_btn_handle=None,
+        error_var=None,
+        quality=80,
+    )
